--- frontend/helpers/findEvent.tsx
@@ -1,4 +1,0 @@
-import { eventData } from "@/pages/_app";
-export default function findEvent (index: number) {
-  return eventData[index]
-}
