+++ frontend/helpers/index.tsx
@@ -1,2 +1,3 @@
 export * from '@/helpers/a11yProps';
 export * from '@/helpers/filterHelpers';
+export * from '@/helpers/finderHelpers';
